requirements.txt
langchain==0.3.14
langchain-openai==0.3.1
langchain-chroma==0.2.1
langsmith==0.3.6
langgraph==0.2.60
chromadb==0.6.3
openai==1.59.9
python-dotenv==1.0.1
tiktoken==0.8.0
deepeval>=3.8.9
guardrails-ai>=0.9.1
presidio-analyzer>=2.2
presidio-anonymizer>=2.2
